fix: remove link statistics and ga served events (#243)
* refactor(stats): rework sequelize queries

LinkStatisticsRepository updates click stats by finding the instance,
then incrementing. This may result in larger transactions than needed,
given that the increment can be done in a single statement. Rework
the implementation so that we update, and if not present, we insert.
We will also do this in many small transactions.

Attempts to implement using `Model.upsert()` were not successful, due
to Sequelize creating temporary functions which fail to reference the
target table.

- interface - rework jsdoc, change return type to boolean
- impl - use `Model.update()` and `Sequelize.literal`, if not found,
  `Model.insert()` or throw not found error
- Drop the constraint of doing stats updates in a single txn

* perf(db): allow db pool max to be controlled, default 20

* perf(stats): remove ga logging for served event

* fix(stats): remove link stats logging except for total clicks

* perf(stats): update default db pool size

Co-authored-by: Yuan Ruo <liangyuanruo@gmail.com>

Co-authored-by: LoneRifle <LoneRifle@users.noreply.github.com>
Co-authored-by: Yuan Ruo <liangyuanruo@gmail.com>