fix(nodemailer): use default value for maxMessages
`maxMessages` is set at `Infinity`, which encourages long-living SMTP
connections. This may not work well with services like AWS SES' SMTP
relays, whose machines are ephemeral and may be culled without notice.

Rely on nodemailer's default of 100 messages per [1]. This allows us
to discard and replace connections to better ensure we are talking to
a live connection. Problems with connection leaks were fixed in
nodemailer 6

References:
[1] - https://nodemailer.com/smtp/pooled/