fix(safebrowsing): correct boolean condition for isThreat
`isThreat()` was evaluated as `safeBrowsingLogOnly && matches`.
This is incorrect as we want the result to always be false when
`safeBrowsingLogOnly`, and to be influenced by matches when
`!safeBrowsingLogOnly`. Correct this typo so that we can correctly
report SafeBrowsing results to the user.

Provide coverage test as part of the work done.