fix(searchable-toggle): update toaster message to reflect changes
Reason is that even though the state of visibility has changed, the toaster visually hasn’t, which may be confusing to users. Since the changed visibility state is clearly stated in the edit link panel (Colour changing text), the toaster should just be a confirmation that a change took place.