refactor(redirect): make into self-contained module
As part of a wider effort to reorganise the codebase around
domain-specific modules, consolidate all redirect-related
objects into a redirect module, under the `redirect/` dir.

Under this new scheme, the `index.ts` of the module will export
by default something that is mountable into express, either a router
or a controller. The index file will also export relevant controllers
and types.

Also take the trouble to dispense with interfaces. By and large, we
no longer have to use interfaces to specify a contract in TypeScript
[1]. Moreover, because we don't actually have multiple implementations
of the interfaces we have declared, we can just use classes, and then
mock the class using the `Required<T>` utility type in TypeScript [2],
or - in this case - using jest.Mocked

- Move routing, controller, services, types to `redirect/`
- GaLogger -> AnalyticsLogger
- Remove all interfaces for redirection

TODO: Move tests to under src, move link statistics logging to
AnalyticsLoggerService

References:
[1]: https://docs.microsoft.com/en-gb/archive/blogs/kcwalina/api-design-myth-interface-as-contract
[2]: https://stackoverflow.com/questions/31876947/confused-about-the-interface-and-class-coding-guidelines-for-typescript