refactor(modules): no interfaces for threat scan services
Interfaces were declared for threat scanners for initial requirements
that multiple services may be used to scan for threats. An interface
was needed then to ensure consistency between the services. Given
however that we have decided on Cloudmersive and Google Safe Browsing
for file and url threats respectively, these interfaces are no longer
needed.

- Remove interfaces from threat module
- Remove dead code in CloudmersiveScanService